Reorganize exports and include flow export file
