+++ flow-exports.txt
@@ -1,0 +1,2 @@
+// @flow
+export * from '../src/index.js'
